Split LGraphGroup out to TS file
